python-simple-rdb-sb/requirements.txt
py-postgresql
pyservicebinding
Flask
gunicorn